* argument types to GestureHandler().checkActionBar() converted from float to int

+++ app/src/main/java/jetsetilly/mandelbrot/Gestures/GestureHandler.java
@@ -1,7 +1,7 @@
 package jetsetilly.mandelbrot.Gestures;
 
 public interface GestureHandler {
-    void checkActionBar(float x, float y, boolean allow_show);
+    void checkActionBar(int x, int y, boolean allow_show);
     void scroll(int x, int y);
     void finishManualGesture();
     void autoZoom(int offset_x, int offset_y, boolean zoom_out);
